feat : 자전거 주행 칼로리 계산 함수 추가

+++ app/src/main/java/com/first/ridingpartnerinsiheung/data/RidingData.kt
@@ -6,5 +6,5 @@ data class RidingData(
     var sumDistance : Double,
     var averSpeed : Double,
     var timer: Int,
-    var kcal: Int
+    var kcal: Double
 )
